fix:  perf, observ and refactor
