+++ .github/workflows/ci.yml
@@ -18,10 +18,36 @@ jobs:
       - name: Check EditorConfig compliance
         uses: editorconfig-checker/action-editorconfig-checker@main
 
+  # ─── Frontend Type Check (required gate) ───────────────────────────────────
+  frontend-typecheck:
+    name: Frontend Type Check
+    runs-on: ubuntu-latest
+
+    defaults:
+      run:
+        working-directory: frontend
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version-file: ".nvmrc"
+          cache: "npm"
+          cache-dependency-path: frontend/package-lock.json
+
+      - name: Install dependencies
+        run: npm ci
+
+      - name: Type check
+        run: npm run type-check
+
   # ─── Frontend ──────────────────────────────────────────────────────────────
   frontend:
     name: Frontend (Next.js)
     runs-on: ubuntu-latest
+    needs: frontend-typecheck
 
     defaults:
       run:
@@ -46,9 +72,6 @@ jobs:
       - name: Lint
         run: npm run lint
 
-      - name: Type check
-        run: npm run type-check
-
       - name: Unit tests
         run: npm test
 
@@ -58,6 +81,10 @@ jobs:
           NEXT_PUBLIC_STELLAR_NETWORK: testnet
           NEXT_PUBLIC_HORIZON_URL: https://horizon-testnet.stellar.org
           NEXT_PUBLIC_API_URL: http://localhost:4000
+          SENTRY_AUTH_TOKEN: ${{ secrets.SENTRY_AUTH_TOKEN }}
+          SENTRY_ORG: ${{ secrets.SENTRY_ORG }}
+          SENTRY_PROJECT: ${{ secrets.SENTRY_PROJECT }}
+          SENTRY_RELEASE: ${{ github.sha }}
 
       - name: Build Storybook (dev-only docs bundle)
         run: npm run build-storybook
@@ -145,6 +172,7 @@ jobs:
   e2e:
     name: E2E Tests (Playwright)
     runs-on: ubuntu-latest
+    needs: frontend-typecheck
 
     defaults:
       run:
