+++ frontend/components/dashboard/index.ts
@@ -1,0 +1,8 @@
+export { BubbleNotification } from "./BubbleNotification";
+export { DraggableWidget } from "./DraggableWidget";
+export { StatsCard, formatStatsXLM } from "./StatsCard";
+export { PaymentStatsWidget } from "./PaymentStatsWidget";
+export type { PaymentStats } from "./PaymentStatsWidget";
+export { MonthlySpendingChart, ThirtyDayVolumeChart } from "./VolumeCharts";
+export { TopRecipientsWidget } from "./TopRecipientsWidget";
+export { BalanceSparkline } from "./BalanceSparkline";
