+++ frontend/pages/dashboard.tsx
@@ -63,17 +63,15 @@ const AIPaymentAssistant = dynamic(() => import("../components/AIPaymentAssistan
 });
 
 import {
-  ResponsiveContainer,
-  BarChart,
-  Bar,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-} from "recharts";
-
-
-import ExternalPaymentBanner from "@/components/ExternalPaymentBanner";
+  DraggableWidget,
+  BubbleNotification,
+  PaymentStatsWidget,
+  MonthlySpendingChart,
+  ThirtyDayVolumeChart,
+  TopRecipientsWidget,
+  BalanceSparkline,
+  PaymentStats,
+} from "@/components/dashboard";
 import PaymentRequestGenerator from "@/pages/PaymentRequestGenerator";
 
 import {
@@ -99,23 +97,6 @@ import { useOnboarding } from "@/hooks/useOnboarding";
 
 interface DashboardProps {
   stellarURI?: URIParseResult | null;
-}
-
-interface PaymentStats {
-  publicKey: string;
-  totalSentXLM: string;
-  totalReceivedXLM: string;
-  sentCount: number;
-  receivedCount: number;
-  totalTransactions: number;
-  comparison?: {
-    thisWeekCount: number;
-    lastWeekCount: number;
-    countChangePercent: number;
-    thisWeekVolume: string;
-    lastWeekVolume: string;
-    volumeChangePercent: number;
-  };
 }
 
 interface CachedBalanceSnapshot {
@@ -1514,443 +1495,11 @@ export default function Dashboard({ stellarURI }: DashboardProps) {
   );
 }
 
-function DraggableWidget({
-  id,
-  dragHandleLabel,
-  isDragging,
-  isDragOver,
-  onDragStart,
-  onDragOver,
-  onDragLeave,
-  onDrop,
-  onDragEnd,
-  children,
-}: {
-  id: string;
-  dragHandleLabel: string;
-  isDragging: boolean;
-  isDragOver: boolean;
-  onDragStart: (e: React.DragEvent) => void;
-  onDragOver: (e: React.DragEvent) => void;
-  onDragLeave: () => void;
-  onDrop: (e: React.DragEvent) => void;
-  onDragEnd: () => void;
-  children: React.ReactNode;
-}) {
-  return (
-    <div
-      onDragOver={onDragOver}
-      onDragLeave={onDragLeave}
-      onDrop={onDrop}
-      data-widget-id={id}
-      className={`relative group rounded-2xl transition-opacity ${isDragging ? "opacity-40" : ""} ${
-        isDragOver ? "ring-2 ring-stellar-400/60 ring-offset-2 ring-offset-cosmos-950 rounded-2xl" : ""
-      }`}
-    >
-      <button
-        type="button"
-        draggable
-        onDragStart={onDragStart}
-        onDragEnd={onDragEnd}
-        title="Drag to reorder"
-        aria-label={`Drag to reorder ${dragHandleLabel}`}
-        className="absolute -top-2 right-2 z-10 flex items-center justify-center cursor-grab active:cursor-grabbing opacity-0 group-hover:opacity-100 focus:opacity-100 transition-opacity bg-white/10 hover:bg-white/20 border border-white/10 rounded-full p-1.5"
-      >
-        <GripIcon className="w-4 h-4 text-slate-300" />
-      </button>
-      {children}
-    </div>
-  );
-}
-
-function GripIcon({ className }: { className?: string }) {
-  return (
-    <svg className={className} viewBox="0 0 24 24" fill="currentColor">
-      <circle cx="9" cy="6" r="1.5" />
-      <circle cx="15" cy="6" r="1.5" />
-      <circle cx="9" cy="12" r="1.5" />
-      <circle cx="15" cy="12" r="1.5" />
-      <circle cx="9" cy="18" r="1.5" />
-      <circle cx="15" cy="18" r="1.5" />
-    </svg>
-  );
-}
-
-function BubbleNotification({ message, visible }: { message: string; visible: boolean }) {
-  return (
-    <div
-      className={`fixed top-4 left-1/2 transform -translate-x-1/2 z-50 transition-all duration-500 ${
-        visible ? 'opacity-100 translate-y-0' : 'opacity-0 -translate-y-full'
-      }`}
-    >
-      <div className="bg-stellar-500 text-white px-4 py-2 rounded-lg shadow-lg max-w-xs">
-        <p className="text-sm whitespace-nowrap overflow-hidden text-ellipsis">{message}</p>
-      </div>
-    </div>
-  );
-}
-
-function PaymentStatsWidget({
-  stats,
-  loading,
-  error,
-  onRetry,
-}: {
-  stats: PaymentStats | null;
-  loading: boolean;
-  error: string | null;
-  onRetry: () => void;
-}) {
-  if (loading) {
-    return (
-      <section
-        className="grid grid-cols-1 gap-4 sm:grid-cols-3 mb-6"
-        aria-label="Payment stats loading"
-      >
-        <span className="sr-only">Loading payment stats</span>
-        {[0, 1, 2].map((index) => (
-          <div
-            key={index}
-            className="card border-white/10 bg-white/[0.03] animate-pulse"
-          >
-            <div className="h-3 w-24 rounded bg-white/10 mb-3" />
-            <div className="h-8 w-32 rounded bg-white/10 mb-2" />
-            <div className="h-3 w-20 rounded bg-white/10" />
-          </div>
-        ))}
-      </section>
-    );
-  }
-
-  if (error) {
-    return (
-      <section className="card mb-6 border-red-500/20 bg-red-500/5">
-        <div className="flex flex-col gap-3 sm:flex-row sm:items-center sm:justify-between">
-          <div>
-            <p className="text-sm font-semibold text-white">Payment summary</p>
-            <p className="text-sm text-red-300">{error}</p>
-          </div>
-          <button onClick={onRetry} className="btn-secondary text-sm px-4 py-2">
-            Retry
-          </button>
-        </div>
-      </section>
-    );
-  }
-
-  if (!stats) return null;
-
-  const countDelta = stats.comparison?.countChangePercent;
-  const volumeDelta = stats.comparison?.volumeChangePercent;
-
-  return (
-    <section className="grid grid-cols-1 gap-4 sm:grid-cols-3 mb-6">
-      <StatsCard
-        label="Total Sent"
-        value={formatStatsXLM(stats.totalSentXLM)}
-        helper={`${stats.sentCount} outgoing payment${stats.sentCount === 1 ? "" : "s"}`}
-        delta={volumeDelta}
-        deltaType={typeof volumeDelta === "number" ? (volumeDelta > 0 ? "positive" : volumeDelta < 0 ? "negative" : "neutral") : undefined}
-      />
-      <StatsCard
-        label="Total Received"
-        value={formatStatsXLM(stats.totalReceivedXLM, "received")}
-        helper={`${stats.receivedCount} incoming payment${stats.receivedCount === 1 ? "" : "s"}`}
-      />
-      <StatsCard
-        label="Transactions"
-        value={stats.totalTransactions.toLocaleString("en-US")}
-        helper="Across sent and received activity"
-        delta={countDelta}
-        deltaType={typeof countDelta === "number" ? (countDelta > 0 ? "positive" : countDelta < 0 ? "negative" : "neutral") : undefined}
-      />
-    </section>
-  );
-}
-
-function MonthlySpendingChart({
-  data,
-  loading,
-  onBarClick,
-}: {
-  data: any[];
-  loading: boolean;
-  onBarClick: (data: any) => void;
-}) {
-  if (loading && data.length === 0) {
-    return (
-      <div className="card mb-6 h-[350px] animate-pulse bg-white/[0.03] border-white/10" />
-    );
-  }
-
-  return (
-    <div className="card mb-6 overflow-hidden">
-      <h2 className="font-display text-lg font-semibold text-white mb-6">
-        Monthly Spending (XLM)
-      </h2>
-      <div className="h-[250px] w-full">
-        <ResponsiveContainer width="100%" height="100%">
-          <BarChart
-            data={data}
-            onClick={(state: any) =>
-              state &&
-              state.activePayload &&
-              onBarClick(state.activePayload[0].payload)
-            }
-
-          >
-            <CartesianGrid strokeDasharray="3 3" stroke="#334155" vertical={false} />
-            <XAxis
-              dataKey="month"
-              axisLine={false}
-              tickLine={false}
-              tick={{ fill: "#94a3b8", fontSize: 12 }}
-            />
-            <YAxis
-              axisLine={false}
-              tickLine={false}
-              tick={{ fill: "#94a3b8", fontSize: 12 }}
-              tickFormatter={(value: any) => `${value}`}
-            />
-            <Tooltip
-              cursor={{ fill: "rgba(255, 255, 255, 0.05)" }}
-              contentStyle={{
-                backgroundColor: "#0f172a",
-                border: "1px solid rgba(255, 255, 255, 0.1)",
-                borderRadius: "8px",
-              }}
-              itemStyle={{ color: "#38bdf8" }}
-            />
-            <Bar dataKey="sent" fill="#38bdf8" radius={[4, 4, 0, 0]} />
-          </BarChart>
-        </ResponsiveContainer>
-      </div>
-    </div>
-  );
-}
-
-function ThirtyDayVolumeChart({ data, loading }: { data: any[]; loading: boolean }) {
-  if (loading && data.length === 0) {
-    return <div className="card mb-6 h-[280px] animate-pulse bg-white/[0.03] border-white/10" />;
-  }
-  const visibleData = data.filter((_: any, i: number) => i % 5 === 0 || i === data.length - 1);
-  return (
-    <div className="card mb-6 overflow-hidden">
-      <h2 className="font-display text-lg font-semibold text-white mb-6">30-Day Volume (XLM)</h2>
-      <div className="h-[220px] w-full">
-        <ResponsiveContainer width="100%" height="100%">
-          <BarChart data={data}>
-            <CartesianGrid strokeDasharray="3 3" stroke="#334155" vertical={false} />
-            <XAxis
-              dataKey="day"
-              axisLine={false}
-              tickLine={false}
-              tick={{ fill: "#94a3b8", fontSize: 11 }}
-              ticks={visibleData.map((d: any) => d.day)}
-              interval="preserveStartEnd"
-            />
-            <YAxis
-              axisLine={false}
-              tickLine={false}
-              tick={{ fill: "#94a3b8", fontSize: 11 }}
-            />
-            <Tooltip
-              cursor={{ fill: "rgba(255,255,255,0.05)" }}
-              contentStyle={{ backgroundColor: "#0f172a", border: "1px solid rgba(255,255,255,0.1)", borderRadius: "8px" }}
-              itemStyle={{ color: "#38bdf8" }}
-            />
-            <Bar dataKey="sent" fill="#38bdf8" name="Sent" radius={[3, 3, 0, 0]} />
-            <Bar dataKey="received" fill="#34d399" name="Received" radius={[3, 3, 0, 0]} />
-          </BarChart>
-        </ResponsiveContainer>
-      </div>
-    </div>
-  );
-}
-
-function TopRecipientsWidget({
-  recipients,
-  loading,
-}: {
-  recipients: Array<{ address: string; totalXLMSent: string }>;
-  loading: boolean;
-}) {
-  return (
-    <div className="card">
-      <h2 className="font-display text-lg font-semibold text-white mb-4">Top Recipients</h2>
-      {loading ? (
-        <div className="space-y-3">
-          {[1, 2, 3, 4, 5].map((i) => (
-            <div key={i} className="h-10 bg-white/5 rounded-lg animate-pulse" />
-          ))}
-        </div>
-      ) : recipients.length === 0 ? (
-        <p className="text-sm text-slate-400">No sent payments yet.</p>
-      ) : (
-        <ol className="space-y-2">
-          {recipients.map((r, idx) => (
-            <li key={r.address} className="flex items-center justify-between gap-3 p-2 rounded-lg bg-white/[0.02] border border-white/5">
-              <div className="flex items-center gap-3">
-                <span className="text-xs font-bold text-stellar-400 w-5 text-center">{idx + 1}</span>
-                <span className="font-mono text-sm text-slate-200">{shortenAddress(r.address)}</span>
-              </div>
-              <span className="text-sm font-semibold text-white">{parseFloat(r.totalXLMSent).toFixed(2)} XLM</span>
-            </li>
-          ))}
-        </ol>
-      )}
-    </div>
-  );
-}
-
 function DownloadIcon({ className }: { className?: string }) {
   return (
     <svg className={className} fill="none" viewBox="0 0 24 24" stroke="currentColor">
       <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 16v1a3 3 0 003 3h10a3 3 0 003-3v-1m-4-4l-4 4m0 0l-4-4m4 4V4" />
     </svg>
-  );
-}
-
-function StatsCard({
-  label,
-  value,
-  helper,
-  delta,
-  deltaType = "neutral",
-}: {
-  label: string;
-  value: string;
-  helper: string;
-  delta?: number;
-  deltaType?: "positive" | "negative" | "neutral";
-}) {
-  const isPos = deltaType === "positive";
-  const isNeg = deltaType === "negative";
-  const deltaColor = isPos ? "text-emerald-400 bg-emerald-500/10" : isNeg ? "text-rose-400 bg-rose-500/10" : "text-slate-400 bg-slate-500/10";
-  
-  return (
-    <div className="card border-white/10 bg-white/[0.03] relative overflow-hidden flex flex-col justify-between">
-      <div>
-        <div className="flex items-center justify-between mb-2">
-          <p className="label">{label}</p>
-          {typeof delta === "number" && (
-            <span className={`text-[11px] font-semibold px-2 py-0.5 rounded-full ${deltaColor}`}>
-              {delta >= 0 ? "+" : ""}{delta}%
-            </span>
-          )}
-        </div>
-        <p className="font-display text-2xl font-bold text-white">{value}</p>
-      </div>
-      <p className="text-xs text-slate-400 mt-2">{helper}</p>
-    </div>
-  );
-}
-
-function formatStatsXLM(amount: string, suffix = "sent") {
-  const value = parseFloat(amount);
-
-  if (Number.isNaN(value)) return `0.00 XLM ${suffix}`;
-
-  return `${value.toLocaleString("en-US", {
-    minimumFractionDigits: 2,
-    maximumFractionDigits: 7,
-  })} XLM ${suffix}`;
-}
-
-// ─── Sparkline chart ─────────────────────────────────────────────────────────
-
-/**
- * Inline SVG sparkline showing balance change over the last N transactions.
- * Green when the overall trend is upward, red when downward.
- * Hover tooltip shows the running balance delta at each data point.
- */
-function BalanceSparkline({ data }: { data: number[] }) {
-  const W = 160;
-  const H = 40;
-  const PAD = 4;
-
-  const min = Math.min(...data);
-  const max = Math.max(...data);
-  const range = max - min || 1; // avoid division by zero for flat lines
-
-  const points = data.map((v, i) => {
-    const x = PAD + (i / Math.max(data.length - 1, 1)) * (W - PAD * 2);
-    const y = PAD + (1 - (v - min) / range) * (H - PAD * 2);
-    return { x, y, value: v };
-  });
-
-  const polyline = points.map((p) => `${p.x},${p.y}`).join(" ");
-
-  const trend = data[data.length - 1] >= data[0];
-  const color = trend ? "#22c55e" : "#ef4444"; // green-500 / red-500
-  const fillColor = trend ? "rgba(34,197,94,0.12)" : "rgba(239,68,68,0.12)";
-
-  // Closed path for the fill area under the line
-  const fillPath =
-    `M ${points[0].x},${H - PAD} ` +
-    points.map((p) => `L ${p.x},${p.y}`).join(" ") +
-    ` L ${points[points.length - 1].x},${H - PAD} Z`;
-
-  return (
-    <div className="relative inline-block" aria-label="Balance sparkline chart">
-      <svg
-        width={W}
-        height={H}
-        viewBox={`0 0 ${W} ${H}`}
-        role="img"
-        aria-label={`Balance trend: ${trend ? "upward" : "downward"}`}
-      >
-        {/* Fill area */}
-        <path d={fillPath} fill={fillColor} />
-        {/* Line */}
-        <polyline
-          points={polyline}
-          fill="none"
-          stroke={color}
-          strokeWidth="1.5"
-          strokeLinejoin="round"
-          strokeLinecap="round"
-        />
-        {/* Interactive dots with tooltips */}
-        {points.map((p, i) => (
-          <g key={i} className="group">
-            <circle
-              cx={p.x}
-              cy={p.y}
-              r={5}
-              fill="transparent"
-              className="cursor-pointer"
-            />
-            <circle
-              cx={p.x}
-              cy={p.y}
-              r={2.5}
-              fill={color}
-              className="opacity-0 group-hover:opacity-100 transition-opacity"
-            />
-            {/* SVG foreignObject tooltip */}
-            <foreignObject
-              x={Math.min(p.x - 36, W - 76)}
-              y={p.y < H / 2 ? p.y + 6 : p.y - 30}
-              width={72}
-              height={24}
-              className="pointer-events-none opacity-0 group-hover:opacity-100 transition-opacity overflow-visible"
-            >
-              <div
-                className="bg-cosmos-900 border border-white/10 rounded px-1.5 py-0.5 text-xs text-white whitespace-nowrap text-center"
-                style={{ fontSize: "10px" }}
-              >
-                {p.value >= 0 ? "+" : ""}
-                {p.value.toFixed(4)} XLM
-              </div>
-            </foreignObject>
-          </g>
-        ))}
-      </svg>
-      <p className="text-xs mt-0.5" style={{ color, fontSize: "10px" }}>
-        {trend ? "▲ Upward trend" : "▼ Downward trend"}
-      </p>
-    </div>
   );
 }
