+++ frontend/sentry.client.config.ts
@@ -9,9 +9,8 @@ import * as Sentry from "@sentry/nextjs";
 Sentry.init({
   dsn: process.env.NEXT_PUBLIC_SENTRY_DSN,
   environment: process.env.NODE_ENV,
-  // Only enable when a DSN is provided
+  release: process.env.SENTRY_RELEASE || process.env.GITHUB_SHA || "dev",
   enabled: !!process.env.NEXT_PUBLIC_SENTRY_DSN,
   tracesSampleRate: 0.2,
-  // Capture unhandled promise rejections and React error boundaries
   integrations: [Sentry.browserTracingIntegration()],
 });
